`main.py`: `find_cities_coordinates` → `find_Minas_Gerais_cities_coordinates`
I rewrote the function `find_cities_coordinates` into `find_Minas_Gerais_cities_coordinates`.

Benefit: it uses only one dataframe from start to end, narrowing it down step-by-step, instead of extracting data from it and inserting into a dictionary, just to extract again and insert into another dictionary. This was clumsy.

As a result, `find_Minas_Gerais_cities_coordinates` returns a dataframe, not a dictionary. The function that takes its output as input, `find_nearest_gauging_station`, was not yet rewritten to take the dataframe as an input directly. As such, a workaround is used to feed `find_nearest_gauging_station` with something that looks like a dictionary, as it expects to.

The function `find_nearest_gauging_station` must be rewritten in the near future to take advantage of the dataframe returned by `find_Minas_Gerais_cities_coordinates`, in order to eliminate the dictionary-like workaround quoted in the next lines:
    
    # Create a dictionary to hold the coordinates of the nearest gauging station in relation to the sought city:
    nearest_coordinates_dict = {}   
    for city in df_cities_coordinates.index:
        nearest_coordinates_dict[city] = find_nearest_gauging_station({'longitude': df_cities_coordinates.loc[city]['LONGITUDE'], 'latitude': df_cities_coordinates.loc[city]['LATITUDE']}, df_SPEI)